Add basic template for QUIC source element.
This commit adds a gstreamer project to the src directory. The commit also adds the basic template for the QUIC source element "quicsrc".

The source element utilises the LSQuic library. To build, first build the LSQuic library and then set the LSQUIC env var to the path of the top level dir of the LSQuic files. Then run ./autogen followed by make.