Correct error in status report
The status report mentions adding paramters to the mininet switch. This is incorrect, as the parameters will actually be set on links between nodes. This commit corrects this mistake.

This commit also updates the deliverable for this section of the plan to indicate that LSQUIC logs will also be gathered.