Fix logging strings within test framework
Previously, the section of the string which set all elements log level to debug came at the end. This had the unintended effect of reducing the log level for elements such as rtp.

To amend this, the string was rearranged so that the universal logging segment comes at the start.