Add timeout to test framework.
the Quicelements can stall on high loss runs. This can either occur iat the beginning, when no connection is established at all (not a bug), or after a connection has been established (bug).

To account for this I have added a timeout which will kill tests that have stopped progressing.