Run performance test on network before sending any packets
Currently, we begin the experiment after setting up the network. However, this appears to introduce issues. At higher latencies, the initial packets are dropped and their is unintended jitter for the first group of packets.

Performing an iperf performance test before starting the actual test seems to greatly reduce this issue. No packets are dropped and the initial jitter is small an only affects the first three packets at most.