Set udpBw parameter in iperf test
Currently, the udpBw parameter is not set and the default sending rate of 10Mbit/s is used.

As we use higher bandwidths in some tests, it would be good to see a confirmation that these are working properly. This commit utilises our bandwidth test parameter when setting udpBw argument for the iperf test.