Allow longer running tests additional time to complete
It should only take ~60 seconds to stream the test video but it might be interesting to see how long certain implementations take. This commit allows additional time for the tests to complete (up to 3* the run time).

Tests which finish earlier will be detected to prevent this causing a great increase to the run time.