Set tshark to log additional details about quic and tcp.
Previosly, the QUIC pcaps were being stored so that they could later be decrypted using the ssl keys. These took up a significant amout of storage space. Experimentation found that the SSL secrets became available in time to decrypt while the test was running.

This commits alters the QUIC tshark commands such that the SSL secrets are passed to tshark.

Additionally, as we will no longer have access to the pcaps after tests have run, we log some additional details such as stream ID, length and whether a packet contains the fin bit.