Subtract 2 from mtu value when applying rfc 4751
When sending rtp packets over a stream with no message boundaries to delineate the start and end of packets, we add a 2 byte length indicator to the beginning of packets. The rtppayloader does not take this into account, and so we end up with rtp packets larger than the path mtu.

This commit sets the path mtu to 1400-2=1398 for implementations which utilise rfc 4751.