keep_test_running.sh should append rather than overwrite
keep_test_running.sh restarts the tests in case of the process crashing. Currently, the output is written to a stdout and stderr file. If a restart occurs, the output for the initial part of the run will be lost.

This commit changes the redirect such that we append to the output file, rather than overwriting.