Monitor client process when waiting for pipeline execution to end
Previously, we would monitor the server gst-launch process and kill the other threads when this process finished or after a timeout has passed. However, it makes more sense to have the client inidicate end of execution, since the server side pipeline will always finish before the client. This commit swaps the controlling threads.

This commit also replaces reference to tcpdump with "tshark" as the packet analyser we use is now tshark