Updated stack latency analysis for QUIC streams.
Now that we have QUIC implementations which may split rtp packets across different quic packets, we need to update our analysis scripts to handle this and reconstruct the packets.

This commit adds new functions for calculating the arrival and availability times for rtp packets sent over QUIC_SS and QUIC_GOP.

This commit also fixes some bugs which were present in the tcp functions.

Additionally, the commit renames some functions for clarity.

Finally, this commit updates the analyser.py script to use the new functions when appropriate and adds the abaility to catch and print exceptions. This allows analysis to continue if any error occurs.