Fixed issue where tcp stack latency analysis would fail when a packet was lost.
Fixed an issue with the stack latency analysis for TCP. It now ignores retransmission lines on the server side, as the rtp dissector does not work for retransmissions. We are only interested in the timestamp of the original transmission anyway, so this is not an issue. I have also added the payload data to the server shark output for tcp. This will be ignored by the analysis script. But if anything goes wrong with the RTP dissector, we will still have the data available.