Guard against fin arriving in seperate packet.
I have modified lsquic such that packets are sent as soon as data is written. This prevents unnessecary delays where lsquic would wait a significant period of time in order to fill packets past a certain threshold. Similar to setting the TCP_NODELAY option.

As a result of these changes, it is possible for the FIN frame to arrive in a seperate packet from the final STREAM frame for a stream. This can lead to quicsrc getting stuck in a loop as it waits for stream data to arrive.

This commit adds another condition to the while loop which will allow it to detect the end of a stream and exit even if no data has been read.