Update stack_latency_analysis to handle FIN frames
Now that data is sent immediately, the fin frame is often sent in the next packet after the stream is closed. This was causing issues as the FIN frame would usually be coallesced with the STREAM frame of the next stream.

This commit updates stack_latency_analysis.py such that it can handle such a scenario