Move analysis scripts to the data directory.
Rather than existing as part of the test framework, it is more appropriate for the analysis scripts to reside in the data directory.

The test runner would run the analysis scripts at the end, but would no longer have access to the analyser.py script. I have removed the analyser call for final submission.