Do not use defaults specific to my local machine
The QUIC elements had been using a default path for their logging and SSL certificate. This will not work when installed on another machine so the elements now only accept commandlienargs for the Certificate, log file and SSL key.

generate_cert.sh will now be used at the beginning of a test run to generate the certifcate and key in the output directory.